chore(ibmcloud): Move validation tests back to ibmcloud_test
