--- pkg/types/ibmcloud/dns.go
@@ -1,23 +1,0 @@
-// Package ibmcloud contains IBM Cloud-specific structures for installer
-// configuration and management.
-package ibmcloud
-
-// DNSZoneResponse represents a DNS zone response.
-type DNSZoneResponse struct {
-	// Name is the domain name of the zone.
-	Name string
-
-	// ID is the zone's ID.
-	ID string
-
-	// CISInstanceCRN is the IBM Cloud Resource Name for the CIS instance where
-	// the DNS zone is managed.
-	CISInstanceCRN string
-
-	// CISInstanceName is the display name of the CIS instance where the DNS zone
-	// is managed.
-	CISInstanceName string
-
-	// ResourceGroupID is the resource group ID of the CIS instance.
-	ResourceGroupID string
-}
