--- pkg/types/ibmcloud/encryptionkey.go
@@ -1,6 +1,0 @@
-// Package ibmcloud contains IBM Cloud-specific structures for installer
-// configuration and management.
-package ibmcloud
-
-// EncryptionKeyResponse represents an encryption key response.
-type EncryptionKeyResponse struct{}
